refactor(utils): rename dom -> aria
